Use float positions, at least in the client

+++ agario-client/src/main/java/client/model/map/object/MapObject.java
@@ -4,25 +4,26 @@ package client.model.map.object;
 import client.model.map.Map;
 import client.model.map.helper.Attributes;
 
-import org.joml.Vector2i;
+import org.joml.Vector2f;
 
 /**
  *
  * @author zoli-
+ * @author yzsolt
  */
 public abstract class MapObject {
     
     protected Map map;
-    protected Vector2i position;
+    protected Vector2f position;
     protected Attributes attributes;
     
-     public MapObject(Map map, Vector2i position, Attributes attributes) {
+     public MapObject(Map map, Vector2f position, Attributes attributes) {
         this.map = map;
         this.position = position;
         this.attributes = attributes;
      }
     
-    public Vector2i getPosition(){
+    public Vector2f getPosition(){
         return this.position;
     }
